// web/src/types/enums/Modelos.ts
export enum MODELOS {
  PACIENTESPOSITIVOS = "Pacientes positivos",
  PACIENTESNEGATIVOS = "Pacientes negativos",
  PACIENTESTODOS = "Pacientes todos",
  PROVINCIACARTILLA = "Provincia cartilla",
  RESULTADOSPENDIENTES = "Resultados Pendientes",
  RESULTADOSPENDIENTESENVIO = "Resultado Pendiente Envio",
  RESULTADOSTODO = "Resultado Todo",
  RESULTADO = "Resultado",
  CIUDAD = "Ciudades",
  CARTILLA = "Cartilla",
  LABORATORIO = "Laboratorio",
  CENTRO = "Centros",
  MUNICIPIO = "Munipios",
  PROVINCIA = "Provincias",
  CONTACTOS = "Contactos",
  CATEGORÍAS_DE_CONTACTO = "Categorías de contacto",
  CUENTAS = "Cuentas",
  LIBRO_MAYOR = "Libro Mayor",
  ASIENTOS = "Asientos",
  DETALLE_DE_ASIENTOS = "Detalle de asientos",
  USUARIOS = "Usuarios",
  GRUPOS = "Grupos",
  MENUS = "Menús",
  MODELOS = "Modelos",
  PRODUCTOS_BASE = "Productos base",
  PRODUCTOS_VARIANTES = "Productos variantes",
  ATRIBUTO = "Atributo",
  ATRIBUTO_VALOR = "Atributo valor",
  BANCOS = "Libro de banco",
  TIPO_CAMBIO = "Tipo de cambio",
  EMPRESA = "Empresa",
  LIBRO_DIARIO = "Libro diario",
  MONEDA = "Monedas",
  EFECTIVO = "Efectivos",
  RMCATEGORIA = "Categoria",
  RMUNIDADMEDIDA = "Unidad medida",
  BANCO = "Banco",
  CONTACTO_BANCO = "Contacto banco",
  METODO_PAGO = "Metodo pago",
  PRODUCTO_CATEGORIA = "Producto categoria",
  SUMAS_SALDOS = "Sumas y saldos",
  FLUJO_EFECTIVO = "Flujo de efectivo",
  BALANCE_GENERAL = "Balance general",
  TARIFA = "Tarifa",
  TIPO_MOVIMIENTO = "Tipo movimiento",
  TIPO_TERMINAL = "Tipo terminal",
  PVCONFIG = "Config",
  MOVIMIENTO = "Movimiento",
  INGRESOEXTRA = "Ingreso extra",
  SALDOFAVOR = "Saldo favor",
  INVENTARIO_CONTABILIDAD = "Inventario",
  CUENTAS_ACTIVOS = "Cuentas de activos",
  ENTREGA_DE_EFECTIVO = "Entrega de efectivo",
  CAJAS = "Caja",
  REGISTRO_DE_GASTOS = "Registro de gastos",
}